sys/lv2_types.h: an EA check that is well defined in both data models
lv2_ea32_pack checked a pointer's upper half with (u >> 32).  In the
default ILP32 data model uintptr_t is 32 bits, so that is a shift by the
full width of the type - undefined behaviour, a warning in every unit that
includes the header (cell/gcm.h pulls it in), and a hard error under
-Werror=shift-count-overflow.  Found by the EMP team porting its engine.
The check now shifts a uint64_t, which is meaningful in LP64 and simply
always false in ILP32, and the header's prose says what the data models
are: 32-bit pointers by default, 64-bit with -mlp64, and ABI-fixed EA
fields 32-bit in both.

Guard: tests/sdk/lv2-ea32-test.{c,sh} compiles the header in C11 and
C++17, ILP32 and LP64, debug and release, under -Wall -Wextra -Werror with
the PPU compiler when present, and runs host round-trips of the EA
boundary values (0, 1, 0x7fffffff, 0x80000000, 0xffffffff) plus, on a
wide-pointer host, the debug trap and release low-bits behaviour.  CI also
compiles it freestanding with -m32, which catches the ILP32 shift without
a PPU toolchain.  The unmodified header fails both.

+++ sdk/include/sys/lv2_types.h
@@ -1,10 +1,10 @@
 /*
  * sys/lv2_types.h — CellOS Lv-2 ABI-level type primitives.
  *
- * The Lv-2 ABI is 64-bit but has a 32-bit userland effective-address
- * (EA) space. The public API exposes 64-bit native C pointers
- * (`void *`, `T *`) for application use, but a small number of ABI-
- * fixed fields carry a 32-bit EA as an explicit `uint32_t`. This
+ * The PPU uses ELF64 with a 32-bit userland effective-address (EA)
+ * space. Native C pointers (`void *`, `T *`) are 32-bit by default
+ * (ILP32), or 64-bit with -mlp64. ABI-fixed EA fields remain 32-bit
+ * in either data model and use an explicit `uint32_t`. This
  * header provides the typedef and conversion helpers that make that
  * distinction explicit, so future code neither loses upper bits by
  * accident nor conflates the two categories.
@@ -19,9 +19,10 @@
  *                      crash_dump_param_addr, any value written to
  *                      an OPD entry-point slot.
  *
- *   void * / T *     — everything else. User-visible pointers in
- *                      public struct fields (CellGcmConfig, etc.)
- *                      are native 64-bit pointers, per Lv-2 ABI.
+ *   void * / T *     — ordinary in-process pointers, whose width
+ *                      follows the selected C data model. Fields
+ *                      crossing the SPRX boundary must preserve
+ *                      their fixed ABI width (see the spec above).
  *
  * Where NOT to use:
  *   Do not use lv2_ea32_t for general-purpose pointer storage.
@@ -46,7 +47,8 @@ typedef uint32_t lv2_ea32_t;
 
 /*
  * Pack a native pointer into a 32-bit EA. In debug builds (non-NDEBUG),
- * asserts that the pointer fits in 32 bits — i.e. its upper half is zero.
+ * traps if the pointer does not fit in 32 bits. This check is vacuous
+ * for ILP32 and rejects nonzero upper bits for LP64.
  * Lv-2 userland EAs always fit in 32 bits; a failing assert means a
  * kernel-returned pointer or a cross-ABI value is being narrowed incorrectly.
  */
@@ -57,7 +59,7 @@ static inline lv2_ea32_t lv2_ea32_pack(const void *p)
     /* Fires on any value whose upper 32 bits aren't zero. Represents a
      * real bug in the caller — a non-userland pointer being stored in
      * a 32-bit ABI slot. */
-    if ((u >> 32) != 0) {
+    if (((uint64_t)u >> 32) != 0) {
         /* Pull in assert.h only when actually needed; keep the hot path
          * include-light. Using __builtin_trap keeps us freestanding. */
         __builtin_trap();
